Refractor augmentations initial commit
